+++ scripts/3_Game/LogZ/DTO/Root.c
@@ -13,7 +13,7 @@ class LogZ_DTO_Root
 	int ts; // UTC time (epoch seconds)
 	int uptime_ms; // server uptime (ms)
 	int world_time; // game world time (epoch seconds)
-	int instance; // instance id (or GamePort or QueryPort as fallback)
+	string instance; // instance id (or GamePort or QueryPort as fallback)
 	string level; // LogZ_Levels level
 	string schema; // LogZ schema version
 	string world; // game world name
@@ -25,13 +25,20 @@ class LogZ_DTO_Root
 	*/
 	void LogZ_DTO_Root(LogZ_Level lvl, string message, LogZ_Event eventType)
 	{
+		if (!LogZ_Config.IsLoaded())
+			return;
+
+		LogZ_ConfigDTO cfg = LogZ_Config.Get();
+		if (!cfg)
+			return;
+
 		ts = LogZ_Time.EpochSecondsUTC();
 		uptime_ms = g_Game.GetTime();
 		world_time = LogZ_Time.GameEpochSeconds();
-		instance = LogZ_Config.GetInstanceID();
+		instance = cfg.settings.instance_id;
 		level = LogZ_Levels.ToString(lvl);
-		schema = LogZ_Config.SCHEMA_VERSION;
-		world = LogZ_Config.GetWorldName();
+		schema = LogZ_Constants.SCHEMA_VERSION;
+		world = cfg.geo.world_name;
 		event_type = LogZ_Events.ToString(eventType);
 		msg = message;
 	}